# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateStickerRequest:
    properties:
      color:
        enum:
        - yellow
        - blue
        - green
        - red
        - purple
        - orange
        - gray
        type: string
      content:
        type: string
      due_date:
        type: string
      height:
        type: integer
      position_x:
        type: integer
      position_y:
        type: integer
      priority:
        enum:
        - low
        - medium
        - high
        - urgent
        type: string
      title:
        type: string
      width:
        type: integer
    required:
    - content
    type: object
...
  dto.StickerResponse:
    properties:
      author:
        $ref: '#/definitions/dto.UserResponse'
      board_id:
        type: string
      color:
        type: string
      content:
        type: string
      created_at:
        type: string
      due_date:
        type: string
      height:
        type: integer
      id:
        type: string
      is_completed:
        type: boolean
      position_x:
        type: integer
      position_y:
        type: integer
      priority:
        type: string
      title:
        type: string
      updated_at:
        type: string
      width:
        type: integer
    type: object
...
  dto.UpdateStickerPositionRequest:
    properties:
      position_x:
        type: integer
      position_y:
        type: integer
    required:
    - position_x
    - position_y
    type: object
  dto.UpdateStickerRequest:
    properties:
      color:
        enum:
        - yellow
        - blue
        - green
        - red
        - purple
        - orange
        - gray
        type: string
      content:
        type: string
      due_date:
        type: string
      height:
        type: integer
      priority:
        enum:
        - low
        - medium
        - high
        - urgent
        type: string
      title:
        type: string
      width:
        type: integer
    type: object
...
  dto.UserResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      role:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@synergyconnect.com
    name: API Support
  description: Университетская социальная сеть с полным набором функций
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://synergyconnect.com/terms
  title: SynergyConnect API
  version: 1.0.0
paths:
...
  /boards/{boardId}/stickers:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID доски
        in: path
        name: boardId
        required: true
        type: string
      - description: Данные стикера
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStickerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.StickerResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Создать стикер
      tags:
      - stickers
...
  /stickers/{id}:
    delete:
      parameters:
      - description: ID стикера
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Удалить стикер
      tags:
      - stickers
    get:
      parameters:
      - description: ID стикера
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StickerResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Получить стикер
      tags:
      - stickers
    put:
      consumes:
      - application/json
      parameters:
      - description: ID стикера
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStickerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StickerResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Обновить стикер
      tags:
      - stickers
  /stickers/{id}/position:
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID стикера
        in: path
        name: id
        required: true
        type: string
      - description: Новая позиция
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStickerPositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StickerResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Обновить позицию стикера
      tags:
      - stickers
  /stickers/{id}/toggle-complete:
    patch:
      parameters:
      - description: ID стикера
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StickerResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Переключить статус выполнения
      tags:
      - stickers
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"